tests/wpt/web-platform-tests/permissions/permissionsstatus-name.html

<!doctype html>
<meta charset=utf-8>
<title>Test PermissionStatus's name attribute.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>

<script>
promise_test(async () => {
  const result = await navigator.permissions.query({
    name: "geolocation",
  });
  assert_equals(result.name, "geolocation", "Name was geolocation");
});
</script>
