third_party/blink/web_tests/external/wpt/permissions/permissionsstatus-name-expected.txt

This is a testharness.js-based test.
FAIL Test PermissionStatus's name attribute. assert_equals: Name was geolocation expected (string) "geolocation" but got (undefined) undefined
Harness: the test ran to completion.
